CONSOLE-4538: Move location of `ErrorDetailsModal`
