+++ frontend/packages/console-shared/locales/en/console-shared.json
@@ -133,11 +133,6 @@
   "Something wrong happened": "Something wrong happened",
   "An error occurred. Please try again.": "An error occurred. Please try again.",
   "Reload page": "Reload page",
-  "Description:": "Description:",
-  "Component trace:": "Component trace:",
-  "Stack trace:": "Stack trace:",
-  "Show details": "Show details",
-  "Close": "Close",
   "You made changes to this page.": "You made changes to this page.",
   "Click {{submit}} to save changes or {{reset}} to cancel changes.": "Click {{submit}} to save changes or {{reset}} to cancel changes.",
   "Reload": "Reload",
@@ -167,6 +162,7 @@
   "Form view is disabled for this chart because the schema is not available": "Form view is disabled for this chart because the schema is not available",
   "Getting started resources": "Getting started resources",
   "Use our collection of resources to help you get started with the Console.": "Use our collection of resources to help you get started with the Console.",
+  "Close": "Close",
   "Expandable details": "Expandable details",
   "More info": "More info",
   "View all quick starts": "View all quick starts",
@@ -209,6 +205,10 @@
   "Delete": "Delete",
   "This action cannot be undone. All associated Deployments, Routes, Builds, Pipelines, Storage/PVCs, Secrets, and ConfigMaps will be deleted.": "This action cannot be undone. All associated Deployments, Routes, Builds, Pipelines, Storage/PVCs, Secrets, and ConfigMaps will be deleted.",
   "Confirm deletion by typing <1>{{resourceName}}</1> below:": "Confirm deletion by typing <1>{{resourceName}}</1> below:",
+  "Description:": "Description:",
+  "Component trace:": "Component trace:",
+  "Stack trace:": "Stack trace:",
+  "Show details": "Show details",
   "No projects found": "No projects found",
   "No namespaces found": "No namespaces found",
   "No results match the filter criteria.": "No results match the filter criteria.",
