+++ frontend/packages/console-shared/src/components/modals/ErrorDetailsModal.tsx
